requirements.txt
google-generativeai
tiktoken
python-dotenv